+++ src/definitions/base.ts
@@ -1,12 +1,12 @@
-import { Chain } from 'viem'
 import { base as vbase, baseSepolia as vbases } from 'viem/chains'
+import { EcoChain } from '../chain.interface'
 
 /**
  * Base Mainnet chain configuration
  * Extends viem's base configuration with Eco-specific RPC URLs and contract addresses
  * Includes Hyperlane Mailbox contract configuration
  */
-export const base: Chain = {
+export const base: EcoChain = {
   ...vbase,
   rpcUrls: {
     ...vbase.rpcUrls,
@@ -20,13 +20,14 @@ export const base: Chain = {
       address: '0xeA87ae93Fa0019a82A727bfd3eBd1cFCa8f64f1D',
     },
   },
+  isCalderaChain: false,
 }
 /**
  * Base Sepolia testnet chain configuration
  * Extends viem's Base Sepolia configuration with Eco-specific RPC URLs and contract addresses
  * Includes Hyperlane Mailbox contract configuration for the testnet
  */
-export const baseSepolia: Chain = {
+export const baseSepolia: EcoChain = {
   ...vbases,
   rpcUrls: {
     ...vbases.rpcUrls,
@@ -40,4 +41,5 @@ export const baseSepolia: Chain = {
       address: '0x6966b0E55883d49BFB24539356a2f8A673E02039',
     },
   },
+  isCalderaChain: false,
 }