+++ src/definitions/celo.ts
@@ -1,15 +1,16 @@
-import { Chain } from 'viem'
 import { celo as vcelo } from 'viem/chains'
+import { EcoChain } from '../chain.interface'
 
-export const celo: Chain = {
+export const celo: EcoChain = {
   ...vcelo,
   rpcUrls: {
-    default: {
+    ...vcelo.rpcUrls,
+    quicknode: {
       http: [
-        `https://responsive-lingering-yard.celo-mainnet.quiknode.pro/{QUICKNODE_API_KEY}`,
+        'https://cold-polished-violet.celo-mainnet.quiknode.pro/${QUICKNODE_API_KEY}',
       ],
       webSocket: [
-        `wss://responsive-lingering-yard.celo-mainnet.quiknode.pro/{QUICKNODE_API_KEY}`,
+        'wss://cold-polished-violet.celo-mainnet.quiknode.pro/${QUICKNODE_API_KEY}',
       ],
     },
   },
@@ -19,4 +20,5 @@ export const celo: Chain = {
       address: '0x50da3B3907A08a24fe4999F4Dcf337E8dC7954bb',
     },
   },
+  isCalderaChain: false,
 }